+++ src/main/java/com/yeolsimee/moneysaving/app/category/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
-package com.yeolsimee.moneysaving.app.routine.repository;
+package com.yeolsimee.moneysaving.app.category.repository;
 
-import com.yeolsimee.moneysaving.app.routine.entity.Category;
+import com.yeolsimee.moneysaving.app.category.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
