--- src/main/java/com/yeolsimee/moneysaving/app/routine/entity/RoutineCheckYN.java
@@ -1,6 +1,0 @@
-package com.yeolsimee.moneysaving.app.routine.entity;
-
-public enum RoutineCheckYN {
-    Y,
-    N
-}
