+++ src/main/java/com/yeolsimee/moneysaving/app/routineday/entity/RoutineCheckYN.java
@@ -1,0 +1,6 @@
+package com.yeolsimee.moneysaving.app.routineday.entity;
+
+public enum RoutineCheckYN {
+    Y,
+    N
+}
