+++ src/main/java/com/yeolsimee/moneysaving/app/routineday/service/RoutineDayService.java
@@ -1,7 +1,7 @@
-package com.yeolsimee.moneysaving.app.routine.service;
+package com.yeolsimee.moneysaving.app.routineday.service;
 
-import com.yeolsimee.moneysaving.app.routine.entity.RoutineDay;
-import com.yeolsimee.moneysaving.app.routine.repository.RoutineDayRespository;
+import com.yeolsimee.moneysaving.app.routineday.entity.RoutineDay;
+import com.yeolsimee.moneysaving.app.routineday.repository.RoutineDayRespository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
