+++ src/test/java/com/yeolsimee/moneysaving/unit/CategoryServiceTest.java
@@ -1,7 +1,7 @@
 package com.yeolsimee.moneysaving.unit;
 
-import com.yeolsimee.moneysaving.app.routine.entity.Category;
-import com.yeolsimee.moneysaving.app.routine.service.CategoryService;
+import com.yeolsimee.moneysaving.app.category.entity.Category;
+import com.yeolsimee.moneysaving.app.category.service.CategoryService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
